fix: resolve compile errors and add error handling across UI scripts
- Fix BattleResultScreen using non-existent Rewards property (use ExpGained/GoldGained)
- Fix MonsterEntity array index type error (long → int cast for RoleColors)
- Add null guard to MonsterHpBar slider to prevent NullReferenceException
- Remove self-referencing using directive in SessionData
- Add missing Cardmong.Data using in BattleScreen
- Wrap all async void handlers in try/catch so API failures surface as toasts

+++ Assets/_Project/Scripts/Battle/MonsterEntity.cs
@@ -32,7 +32,7 @@ namespace Cardmong.Battle
                 if (spriteRenderer.sprite == null)
                     spriteRenderer.sprite = PlaceholderSprite.Get((int)(userCardId % RoleColors.Length));
 
-                spriteRenderer.color   = RoleColors[userCardId % RoleColors.Length];
+                spriteRenderer.color   = RoleColors[(int)(userCardId % RoleColors.Length)];
                 spriteRenderer.flipX   = flipX;
             }
         }
